requirements.txt
openai
feedparser
PyQt5
streamlit
trafilatura
openai
feedparser
pandas
google-cloud-aiplatform
pandas
